MDEV-25370 Update for portion changes autoincrement key in bi-temp table
According to the standard, the autoincrement column (i.e. *identity
column*) should be advanced each insert implicitly made by
UPDATE/DELETE ... FOR PORTION.

This is very unconvenient use in several notable cases. Concider a
WITHOUT OVERLAPS key with an autoinc column:
id int auto_increment, unique(id, p without overlaps)

An update or delete with FOR PORTION creates a sense that id will remain
unchanged in such case.

This commit makes an exception the autoinc columns that are part of
WITHOUT OVERLAPS keys and infers the following rule for the newly inserted
rows:
=====
Let V_j a set of values for currently updating/deleting row.
Let VI_j a set of values to be inserted.
* Case:
(a) if V_i is an identity column that is also a part of a unique constraint
    that includes <without overlap specification>, then let VI_i be V_i
(b) otherwise, let VI_i be the value deduced as described by the rules of
    ISO/IEC 9075-2, clauses 15.7 Effect of deleting rows from base tables
    and 15.13 Effect of replacing rows in base tables.
=====

Note that this also infers that if an identity column is not a part of
any unique constraint that includes <without overlap specification>, then
its value will be DEFAULT and, that is, auto-incremented.

Apart from WITHOUT OVERLAPS there is also another notable case, referred
by the reporter - a unique key that has an autoincrement column and a field
from the period specification:
  id int auto_increment, unique(id, s), period for p(s, e)

for this case, no exception is made, and the autoincrementing rules will be
proceeded accordung to the standard (i.e. the value will be advanced on
implicit inserts).